+++ packages/foam-cli/src/utils/index.ts
@@ -1,0 +1,4 @@
+import * as fs from 'fs';
+
+
+export const isValidDirectory = (path: string) => fs.existsSync(path) && fs.lstatSync(path).isDirectory();
